manage battery voltage as float in volts

+++ power_management.h
@@ -64,7 +64,7 @@ typedef struct _pm_callbacks_t {
 void pm_init();
 void pm_set_power_keep(bool value);
 bool pm_get_low_battery();
-uint16_t pm_get_battery_voltage();
+float pm_get_battery_voltage();
 void pm_set_peripheral_power(bool value);
 bool pm_get_peripheral_power();
 bool pm_get_power_sw();
